perf(code-intel): apply O(1) line lookup to the remaining 19 scanner sites (#12884)
Follow-up to #12885, which fixed the 4 sites in code_intelligence/security/
analyzer.py that py-spy implicated. The same per-match rescan —
`content[:match.start()].count("\n") + 1`, O(n*m) and holding the GIL in C with
no yield point — appeared 19 more times across 15 files.

All 19 are inside finditer loops, so all suffer the quadratic behaviour. My
first classification said 5 were single-use and safe to skip; spot-checking two
showed both were in loops whose finditer sat further back than my 8-line
lookback. Re-scanning back to the enclosing def gives 19 of 19 — trusting the
first pass would have left 5 real cases unfixed while reporting the sweep
complete.

The helper moved from code_intelligence/shared/ to utils/. Importing
`code_intelligence.shared.line_index` from api/, code_analysis/ or services/
fails at startup — the code_intelligence package is heavy/stubbed in the
import-smoke environment, so traversing it raises ModuleNotFoundError for the
submodule. tests/test_startup_imports.py caught this on
api.analytics_code_review; a unit test would not have. utils/ was verified
import-clean (no package-level imports). The old path remains as a re-export so
the already-merged in-package import in security/analyzer.py keeps working.

Verification: build-before-use ordering checked at all 19 sites (a mechanical
insert can easily land the build after its use); every changed file parses;
helper equivalence re-confirmed. 362 passed across test_startup_imports +
test_line_index, and 140 passed / 0 failed on the analyz/pattern/code_review/
precommit selection (was 1 failed / 139 before the relocation).

The single flake8 F821 in code_analysis/src/api_consistency_analyzer.py is
pre-existing on Dev_new_gui, confirmed against the baseline file.
