+++ autobot-backend/code_analysis/auto-tools/playwright_sanitizer.py
@@ -13,6 +13,7 @@ Author: AutoBot Playwright Security Specialist
 Version: 1.0.0
 """
 
+from utils.line_index import LineIndex  # #12884
 import hashlib
 import json
 import logging
@@ -183,8 +184,11 @@ class PlaywrightSecurityFixer:
         for pattern_name, pattern_info in patterns.items():
             matches = list(re.finditer(pattern_info["regex"], content, re.IGNORECASE))
 
+            # #12884: build the offset->line map once; the per-match
+            # `content[:start].count()` was O(n*m) and held the GIL.
+            _line_index = LineIndex(content)
             for match in matches:
-                line_num = content[: match.start()].count("\n") + 1
+                line_num = _line_index.line_of(match.start())
                 context_start = max(0, match.start() - 50)
                 context_end = min(len(content), match.end() + 50)
                 context = content[context_start:context_end].strip()
