+++ autobot-backend/code_analysis/scripts/analyze_critical_env_vars.py
@@ -5,6 +5,7 @@
 Focused analysis for critical hardcoded environment variables
 """
 
+from utils.line_index import LineIndex  # #12884
 import re
 from pathlib import Path
 from typing import Any, Dict, List
@@ -116,8 +117,11 @@ class CriticalEnvAnalyzer:
 
             # Issue #510: Handle both compiled Pattern and string
             regex = pattern if hasattr(pattern, "finditer") else re.compile(pattern)
+            # #12884: build the offset->line map once; the per-match
+            # `content[:start].count()` was O(n*m) and held the GIL.
+            _line_index = LineIndex(content)
             for match in regex.finditer(content):
-                line_num = content[: match.start()].count("\n") + 1
+                line_num = _line_index.line_of(match.start())
                 value = match.group(1) if match.groups() else match.group(0)
 
                 # Get context
