+++ autobot-backend/code_analysis/src/frontend_analyzer.py
@@ -5,6 +5,7 @@ Frontend Code Analyzer
 Extends the analysis suite to support JavaScript, TypeScript, Vue, React, and other frontend technologies
 """
 
+from utils.line_index import LineIndex  # #12884
 import asyncio
 import json
 import re
@@ -533,8 +534,11 @@ class FrontendAnalyzer:
                     continue
 
                 for compiled_pattern, description, severity in compiled_list:
+                    # #12884: build the offset->line map once; the per-match
+                    # `content[:start].count()` was O(n*m) and held the GIL.
+                    _line_index = LineIndex(content)
                     for match in compiled_pattern.finditer(content):
-                        line_num = content[: match.start()].count("\n") + 1
+                        line_num = _line_index.line_of(match.start())
 
                         issues.append(
                             FrontendIssue(
