+++ autobot-backend/code_intelligence/code_review_engine.py
@@ -25,6 +25,7 @@ Issue #554: Enhanced with Vector/Redis/LLM infrastructure:
 - Historical review pattern learning via embeddings
 """
 
+from utils.line_index import LineIndex  # #12884
 import re
 import subprocess  # nosec B404 - code review tools require subprocess
 from dataclasses import dataclass, field
@@ -554,8 +555,11 @@ class CodeReviewEngine(_BaseClass):
                 if not (path.suffix == "" and ".py" in pattern_def.file_extensions):
                     continue
 
+            # #12884: build the offset->line map once; the per-match
+            # `content[:start].count()` was O(n*m) and held the GIL.
+            _line_index = LineIndex(content)
             for match in pattern.finditer(content):
-                line_num = content[: match.start()].count("\n") + 1
+                line_num = _line_index.line_of(match.start())
                 comments.append(
                     self._create_pattern_comment(
                         pattern_id,
